ai_service/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
google-generativeai
Pillow
python-multipart
pydantic>=2.0  # IMPORTANT: Cần version 2.x